Virtualizer cleanup (patch 1323346)
git-svn-id: http://code.jaspersoft.com/svn/repos/jasperreports/trunk/jasperreports@902 76e25fc1-081a-0410-bb1b-c96bc70638b7

+++ src/net/sf/jasperreports/engine/base/JRVirtualPrintPage.java
@@ -155,7 +155,7 @@ public class JRVirtualPrintPage extends JRBasePrintPage implements
 			return Integer.toString(System.identityHashCode(printObject)) + "_"
 					+ (printObject.getPages().size()) + "_"
 					+ Integer.toString(counter++) + "_"
-					+ Integer.toString((short) random.nextInt());
+					+ Integer.toString(random.nextInt());
 		}
 	}
